requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
dj-database-url==0.5.0
Django==2.1.1
django-pandas==0.5.1
dlib==19.7.0
face-recognition==1.2.3
face-recognition-models==0.3.0

idna==2.7
kaggle==1.4.6
numpy==1.15.1
opencv-python==3.4.2.17
pandas==0.23.4
Pillow==5.2.0
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
scipy==1.1.0
six==1.11.0
trackml==3.dev0
urllib3==1.23
wincertstore==0.2
